Update README.md
As with WeKan v6.11 requirement of Node.js v14.x, you'd be supposed to activate a newer major version of WeKan dependency. This text update reflects the better practice, to check and be aware of such major changes, as the script isn't intended to check all of your apps for possible conflicts.