Create update_wekan3
Alpha. This version takes into account a change introduced to wekan a while back, which I could not port to run wekan as a service anymore. Still a lot of work to do, like making update calls to wekan code server robust, automatically shutting down a running wekan instance, and restarting it using tmux and the wekan start script.