Now using supervisord to run WeKan as a service.
This version returns to running WeKan as a service with supervisord, after growing tired of restarting WeKan manually (when it did crash for some unknown reason). This reintroduced some old but refactored code and, hence, commented out the code block used to run the startup script in a tmux session. Tested working with upgrading to WeKan v7.60.